+++ pkg/services/object/common.go
@@ -122,13 +122,17 @@ func forwardServerStreamRequestGRPC(ctx context.Context, req any, respStream grp
 	}
 }
 
-func callUnary(ctx context.Context, conn *grpc.ClientConn, method string, req any) (mem.BufferSlice, error) {
-	var respBuf mem.BufferSlice
-
-	err := conn.Invoke(ctx, method, req, &respBuf,
+func callUnaryWithCustomResponse(ctx context.Context, conn *grpc.ClientConn, method string, req any, resp any) error {
+	return conn.Invoke(ctx, method, req, resp,
 		grpc.StaticMethod(),
 		grpc.ForceCodecV2(iprotobuf.BufferedCodec{}),
 	)
+}
+
+func callUnary(ctx context.Context, conn *grpc.ClientConn, method string, req any) (mem.BufferSlice, error) {
+	var respBuf mem.BufferSlice
+
+	err := callUnaryWithCustomResponse(ctx, conn, method, req, &respBuf)
 	if err != nil {
 		return nil, fmt.Errorf("sending the request failed: %w", err)
 	}